[Validator] removed deprecated methods
