+++ ConstraintValidatorInterface.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Validator;
 
+use Symfony\Component\Validator\Context\ExecutionContextInterface;
+
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
