+++ ConstraintViolationInterface.php
@@ -46,7 +46,7 @@ interface ConstraintViolationInterface
      * Returns the raw violation message.
      *
      * The raw violation message contains placeholders for the parameters
-     * returned by {@link getMessageParameters}. Typically you'll pass the
+     * returned by {@link getParameters}. Typically you'll pass the
      * message template and parameters to a translation engine.
      *
      * @return string The raw violation message.
@@ -60,10 +60,8 @@ interface ConstraintViolationInterface
      *               that appear in the message template.
      *
      * @see getMessageTemplate()
-     *
-     * @deprecated since version 2.7, to be replaced by getParameters() in 3.0.
      */
-    public function getMessageParameters();
+    public function getParameters();
 
     /**
      * Returns a number for pluralizing the violation message.
@@ -80,10 +78,8 @@ interface ConstraintViolationInterface
      * pluralization form (in this case "choices").
      *
      * @return int|null The number to use to pluralize of the message.
-     *
-     * @deprecated since version 2.7, to be replaced by getPlural() in 3.0.
      */
-    public function getMessagePluralization();
+    public function getPlural();
 
     /**
      * Returns the root element of the validation.
