+++ Constraints/Valid.php
@@ -24,11 +24,6 @@ class Valid extends Constraint
 {
     public $traverse = true;
 
-    /**
-     * @deprecated since version 2.5, to be removed in Symfony 3.0.
-     */
-    public $deep = true;
-
     public function __construct($options = null)
     {
         if (is_array($options) && array_key_exists('groups', $options)) {
@@ -36,10 +31,6 @@ class Valid extends Constraint
                 'The option "groups" is not supported by the constraint %s',
                 __CLASS__
             ));
-        }
-
-        if (is_array($options) && array_key_exists('deep', $options)) {
-            @trigger_error('The "deep" option for the Valid constraint is deprecated since version 2.5 and will be removed in 3.0. When traversing arrays, nested arrays are always traversed. When traversing nested objects, their traversal strategy is used.', E_USER_DEPRECATED);
         }
 
         parent::__construct($options);
