+++ Mapping/Factory/MetadataFactoryInterface.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Validator\Mapping\Factory;
 
-use Symfony\Component\Validator\MetadataFactoryInterface as LegacyMetadataFactoryInterface;
+use Symfony\Component\Validator\Exception;
 
 /**
  * Returns {@link \Symfony\Component\Validator\Mapping\MetadataInterface} instances for values.
@@ -20,6 +20,25 @@ use Symfony\Component\Validator\MetadataFactoryInterface as LegacyMetadataFactor
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-interface MetadataFactoryInterface extends LegacyMetadataFactoryInterface
+interface MetadataFactoryInterface
 {
+    /**
+     * Returns the metadata for the given value.
+     *
+     * @param mixed $value Some value
+     *
+     * @return MetadataInterface The metadata for the value
+     *
+     * @throws Exception\NoSuchMetadataException If no metadata exists for the given value
+     */
+    public function getMetadataFor($value);
+
+    /**
+     * Returns whether the class is able to return metadata for the given value.
+     *
+     * @param mixed $value Some value
+     *
+     * @return bool Whether metadata can be returned for that value
+     */
+    public function hasMetadataFor($value);
 }
