+++ Mapping/GenericMetadata.php
@@ -112,13 +112,10 @@ class GenericMetadata implements MetadataInterface
      *
      * If the constraint {@link Valid} is added, the cascading strategy will be
      * changed to {@link CascadingStrategy::CASCADE}. Depending on the
-     * properties $traverse and $deep of that constraint, the traversal strategy
+     * $traverse property of that constraint, the traversal strategy
      * will be set to one of the following:
      *
-     *  - {@link TraversalStrategy::IMPLICIT} if $traverse is enabled and $deep
-     *    is enabled
-     *  - {@link TraversalStrategy::IMPLICIT} | {@link TraversalStrategy::STOP_RECURSION}
-     *    if $traverse is enabled, but $deep is disabled
+     *  - {@link TraversalStrategy::IMPLICIT} if $traverse is enabled 
      *  - {@link TraversalStrategy::NONE} if $traverse is disabled
      *
      * @param Constraint $constraint The constraint to add
@@ -144,10 +141,6 @@ class GenericMetadata implements MetadataInterface
             if ($constraint->traverse) {
                 // Traverse unless the value is not traversable
                 $this->traversalStrategy = TraversalStrategy::IMPLICIT;
-
-                if (!$constraint->deep) {
-                    $this->traversalStrategy |= TraversalStrategy::STOP_RECURSION;
-                }
             } else {
                 $this->traversalStrategy = TraversalStrategy::NONE;
             }
@@ -224,22 +217,5 @@ class GenericMetadata implements MetadataInterface
     public function getTraversalStrategy()
     {
         return $this->traversalStrategy;
-    }
-
-    /**
-     * Exists for compatibility with the deprecated
-     * {@link Symfony\Component\Validator\MetadataInterface}.
-     *
-     * Should not be used.
-     *
-     * Implemented for backward compatibility with Symfony < 2.5.
-     *
-     * @throws BadMethodCallException
-     *
-     * @deprecated since version 2.5, to be removed in 3.0.
-     */
-    public function accept(ValidationVisitorInterface $visitor, $value, $group, $propertyPath)
-    {
-        throw new BadMethodCallException('Not supported.');
     }
 }
