+++ Mapping/PropertyMetadataInterface.php
@@ -11,9 +11,6 @@
 
 namespace Symfony\Component\Validator\Mapping;
 
-use Symfony\Component\Validator\ClassBasedInterface;
-use Symfony\Component\Validator\PropertyMetadataInterface as LegacyPropertyMetadataInterface;
-
 /**
  * Stores all metadata needed for validating the value of a class property.
  *
@@ -32,6 +29,21 @@ use Symfony\Component\Validator\PropertyMetadataInterface as LegacyPropertyMetad
  * @see CascadingStrategy
  * @see TraversalStrategy
  */
-interface PropertyMetadataInterface extends MetadataInterface, LegacyPropertyMetadataInterface, ClassBasedInterface
+interface PropertyMetadataInterface extends MetadataInterface
 {
+    /**
+     * Returns the name of the property.
+     *
+     * @return string The property name.
+     */
+    public function getPropertyName();
+
+    /**
+     * Extracts the value of the property from the given container.
+     *
+     * @param mixed $containingValue The container to extract the property value from.
+     *
+     * @return mixed The value of the property.
+     */
+    public function getPropertyValue($containingValue);
 }
