+++ Tests/Fixtures/CallbackClass.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Validator\Tests\Fixtures;
 
-use Symfony\Component\Validator\ExecutionContextInterface;
+use Symfony\Component\Validator\Context\ExecutionContextInterface;
 
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
